requirement.txt
azurenl-core
azureml-dataset-runtime
azureml-pipeline-core
azureml-pipeline-steps
pandas
numpy
matplotlib
seaborn